Add compatibility with newer BORIS annotation format
In recent BORIS versions (v8+?), the export of behavioral events
in a tabular format used as annotations files in SimBA has changed.
This commit thus enables SimBA to import these annotations files
in a new format, while maintaining compatibility with the now-older
format.

This is done by adding a quick version checker seeking for the
presence of a column name found only in the new format, followed by
the adjustment of expected columns headers to match what SimBA is
already expecting.